+++ kgcnn/layers/conv/attention.py
@@ -1,7 +1,0 @@
-from kgcnn.layers.conv.gat_conv import *
-from kgcnn.layers.conv.attentivefp_conv import *
-import logging
-
-# For old reference.
-logging.error(
-    "Module '%s' is deprecated and will be removed in future version. Please find `gat_conv` and `attentivefp_conv`.")