+++ kgcnn/layers/conv/bp2ndgen_conv.py
@@ -1,0 +1,4 @@
+import tensorflow as tf
+ks = tf.keras
+
+